refactor: Move JwtProvider to security package
