--- src/main/java/com/synergy/backend/global/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,0 @@
-package com.synergy.backend.global.jwt;
-
-public class JwtAuthenticationFilter {
-}
