+++ src/main/java/com/synergy/backend/global/security/CustomUserDetailsService.java
@@ -1,0 +1,4 @@
+package com.synergy.backend.global.security;
+
+public class CustomUserDetailsService {
+}
