+++ src/main/java/com/synergy/backend/global/security/JwtAuthenticationFilter.java
@@ -1,0 +1,4 @@
+package com.synergy.backend.global.security;
+
+public class JwtAuthenticationFilter {
+}
